cleanup: refactor benchmark configuration
Make it easier to test the configuration parsing and avoid code
duplication.
