+++ google/cloud/spanner/benchmarks/spanner_client_benchmark_programs.bzl
@@ -14,12 +14,10 @@
 #
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 
-"""Automatically generated source lists for spanner_client_benchmarks - DO NOT EDIT."""
+"""Automatically generated unit tests list - DO NOT EDIT."""
 
-spanner_client_benchmarks_hdrs = [
-    "benchmarks_config.h",
-]
-
-spanner_client_benchmarks_srcs = [
-    "benchmarks_config.cc",
+spanner_client_benchmark_programs = [
+    "benchmarks_config_test.cc",
+    "multiple_rows_cpu_benchmark.cc",
+    "single_row_throughput_benchmark.cc",
 ]
